# api/openapi/platform93.yaml
openapi: 3.1.0
info:
  title: Platform93 API
  version: 0.1.0
  license: { name: Apache-2.0, identifier: Apache-2.0 }
  description: Public contract for fresh Platform93 installations.
servers: [{ url: / }]
security: [{ bearerAuth: [] }]
...
paths:
...
  /v1/applications/{application_id}/auth/methods:
    parameters: [{ $ref: "#/components/parameters/ApplicationID" }]
    post: { operationId: authMethods, tags: [Identity], security: [], requestBody: { $ref: "#/components/requestBodies/Object" }, responses: { "200": { description: Enumeration-safe enabled authentication methods } } }
...
  /v1/applications/{application_id}/auth/logout:
    parameters: [{ $ref: "#/components/parameters/ApplicationID" }]
    post: { operationId: logoutCurrentSession, tags: [Identity], responses: { "204": { description: Current interactive session revoked } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT-or-PAT }
    operatorBearer: { type: http, scheme: bearer, bearerFormat: JWT }
    operatorCookie: { type: apiKey, in: cookie, name: p93_operator_access }
    managementBearer: { type: http, scheme: bearer, bearerFormat: Platform93-management-JWT, description: "Short-lived client-credentials token with actor_type=management_client and audience platform93:control." }
  parameters:
    ApplicationID: { name: application_id, in: path, required: true, schema: { $ref: "#/components/schemas/UUID" } }
...
  requestBodies:
...
    Object: { required: true, content: { application/json: { schema: { type: object } } } }
...
  schemas:
...
    UUID: { type: string, format: uuid }
...